Some reorg, split IAuthenticationContext.cs into multiple files
