+++ src/Altinn.App.Api/Controllers/PartiesController.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using Altinn.App.Core.Configuration;
+using Altinn.App.Core.Features.Auth;
 using Altinn.App.Core.Helpers;
 using Altinn.App.Core.Internal.App;
 using Altinn.App.Core.Internal.Auth;
