+++ src/Altinn.App.Core/Features/Auth/AuthenticationInfo.cs
@@ -1,11 +1,7 @@
 using System.Globalization;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Altinn.App.Core.Configuration;
 using Altinn.App.Core.Helpers;
-using Altinn.App.Core.Internal.App;
-using Altinn.App.Core.Internal.Profile;
-using Altinn.App.Core.Internal.Registers;
 using Altinn.App.Core.Models;
 using Altinn.Platform.Profile.Models;
 using Altinn.Platform.Register.Models;
@@ -13,19 +9,8 @@ using AltinnCore.Authentication.Constants;
 using AltinnCore.Authentication.Utils;
 using Authorization.Platform.Authorization.Models;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Options;
 
-namespace Altinn.App.Core.Internal.Auth;
-
-internal static class AuthenticationContextDI
-{
-    internal static void AddAuthenticationContext(this IServiceCollection services)
-    {
-        services.TryAddSingleton<IAuthenticationContext, AuthenticationContext>();
-    }
-}
+namespace Altinn.App.Core.Features.Auth;
 
 /// <summary>
 /// Contains information about the current logged in client/user.
@@ -585,105 +570,4 @@ public abstract record AuthenticationInfo
         [property: JsonPropertyName("authority")] string Authority,
         [property: JsonPropertyName("ID")] string Id
     );
-}
-
-/// <summary>
-/// Provides access to the current authentication context.
-/// </summary>
-internal interface IAuthenticationContext
-{
-    /// <summary>
-    /// The current authentication info.
-    /// </summary>
-    AuthenticationInfo Current { get; }
-}
-
-internal sealed class AuthenticationContext : IAuthenticationContext
-{
-    private const string ItemsKey = "Internal_AltinnAuthenticationInfo";
-    private readonly IHttpContextAccessor _httpContextAccessor;
-    private readonly IOptionsMonitor<AppSettings> _appSettings;
-    private readonly IOptionsMonitor<GeneralSettings> _generalSettings;
-    private readonly IProfileClient _profileClient;
-    private readonly IAltinnPartyClient _altinnPartyClient;
-    private readonly IAuthorizationClient _authorizationClient;
-    private readonly IAppMetadata _appMetadata;
-
-    public AuthenticationContext(
-        IHttpContextAccessor httpContextAccessor,
-        IOptionsMonitor<AppSettings> appSettings,
-        IOptionsMonitor<GeneralSettings> generalSettings,
-        IProfileClient profileClient,
-        IAltinnPartyClient altinnPartyClient,
-        IAuthorizationClient authorizationClient,
-        IAppMetadata appMetadata
-    )
-    {
-        _httpContextAccessor = httpContextAccessor;
-        _appSettings = appSettings;
-        _generalSettings = generalSettings;
-        _profileClient = profileClient;
-        _altinnPartyClient = altinnPartyClient;
-        _authorizationClient = authorizationClient;
-        _appMetadata = appMetadata;
-    }
-
-    // Currently we're coupling this to the HTTP context directly.
-    // In the future we might want to run work (e.g. service tasks) in the background,
-    // at which point we won't always have a HTTP context available.
-    // At that point we probably want to implement something like an `IExecutionContext`, `IExecutionContextAccessor`
-    // to decouple ourselves from the ASP.NET request context.
-    // TODO: consider removing dependcy on HTTP context
-    private HttpContext _httpContext =>
-        _httpContextAccessor.HttpContext ?? throw new InvalidOperationException("No HTTP context available");
-
-    internal void ResolveCurrent(HttpContext httpContext)
-    {
-        var authInfo = AuthenticationInfo.From(
-            httpContext,
-            _appSettings.CurrentValue.RuntimeCookieName,
-            _generalSettings.CurrentValue.GetAltinnPartyCookieName,
-            _profileClient.GetUserProfile,
-            _altinnPartyClient.GetParty,
-            (string orgNr) => _altinnPartyClient.LookupParty(new PartyLookup { OrgNo = orgNr }),
-            _authorizationClient.GetPartyList,
-            _authorizationClient.ValidateSelectedParty,
-            _authorizationClient.GetUserRoles,
-            _appMetadata.GetApplicationMetadata
-        );
-        httpContext.Items[ItemsKey] = authInfo;
-    }
-
-    public AuthenticationInfo Current
-    {
-        get
-        {
-            var httpContext = _httpContext;
-
-            AuthenticationInfo authInfo;
-            if (!httpContext.Items.TryGetValue(ItemsKey, out var authInfoObj))
-            {
-                authInfo = AuthenticationInfo.From(
-                    httpContext,
-                    _appSettings.CurrentValue.RuntimeCookieName,
-                    _generalSettings.CurrentValue.GetAltinnPartyCookieName,
-                    _profileClient.GetUserProfile,
-                    _altinnPartyClient.GetParty,
-                    (string orgNr) => _altinnPartyClient.LookupParty(new PartyLookup { OrgNo = orgNr }),
-                    _authorizationClient.GetPartyList,
-                    _authorizationClient.ValidateSelectedParty,
-                    _authorizationClient.GetUserRoles,
-                    _appMetadata.GetApplicationMetadata
-                );
-                httpContext.Items[ItemsKey] = authInfo;
-            }
-            else
-            {
-                authInfo =
-                    authInfoObj as AuthenticationInfo
-                    ?? throw new Exception("Unexpected type for authentication info in HTTP context");
-            }
-            return authInfo;
-        }
-    }
 }
